requirements.txt
pandas~=1.1.3
requests~=2.24.0
python-dateutil~=2.8.1
colorama~=0.4.3
matplotlib~=3.3.2